+++ src/main/java/com/raoulvdberge/refinedstorage/gui/grid/view/GridViewFluid.java
@@ -5,26 +5,11 @@ import com.raoulvdberge.refinedstorage.gui.grid.sorting.IGridSorter;
 import com.raoulvdberge.refinedstorage.gui.grid.stack.GridStackFluid;
 import com.raoulvdberge.refinedstorage.gui.grid.stack.IGridStack;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class GridViewFluid extends GridViewBase {
-    private Map<Integer, GridStackFluid> map = new HashMap<>();
-    private List<IGridStack> stacks;
-    private GuiGrid gui;
-    private List<IGridSorter> sorters;
-    private boolean canCraft;
-
     public GridViewFluid(GuiGrid gui, List<IGridSorter> sorters) {
-        this.sorters = sorters;
-        this.gui = gui;
-    }
-
-    @Override
-    public List<IGridStack> getStacks() {
-        return stacks;
+        super(gui, sorters);
     }
 
     @Override
@@ -32,16 +17,16 @@ public class GridViewFluid extends GridViewBase {
         map.clear();
 
         for (IGridStack stack : stacks) {
-            map.put(stack.getHash(), (GridStackFluid) stack);
+            map.put(stack.getHash(), stack);
         }
     }
 
     @Override
     public void postChange(IGridStack stack, int delta) {
-        GridStackFluid existing = map.get(stack.getHash());
+        GridStackFluid existing = (GridStackFluid) map.get(stack.getHash());
 
         if (existing == null) {
-            map.put(stack.getHash(), (GridStackFluid) stack);
+            map.put(stack.getHash(), stack);
         } else {
             if (existing.getStack().amount + delta <= 0) {
                 map.remove(existing.getHash());
@@ -51,30 +36,5 @@ public class GridViewFluid extends GridViewBase {
 
             existing.setTrackerEntry(stack.getTrackerEntry());
         }
-    }
-
-    @Override
-    public void setCanCraft(boolean canCraft) {
-        this.canCraft = canCraft;
-    }
-
-    @Override
-    public boolean canCraft() {
-        return canCraft;
-    }
-
-    @Override
-    public void sort() {
-        List<IGridStack> stacks = new ArrayList<>();
-
-        if (gui.getGrid().isActive()) {
-            stacks.addAll(map.values());
-
-            new Thread(() -> sortAndFilter(gui, stacks, sorters)).start();
-        }
-
-        this.stacks = stacks;
-
-        updateUI(gui);
     }
 }